Add `iterable::Reverse`.
This code comes from arkworks-rs/gemini and is used mostly within tests
in order to efficiently provide streams to both time- and space- efficient
algorithms in order to check consistency of the output.

Despite this code is relevant only in testing environments,
cfg(test) does not have visibility across libraries.
Therefore, I'm opting for exposing it as public API.
One alternative (maybe in the future?) is to use a new feature flag "test"
to expose testing utilities, but feels overkill for this single one.